move checkpoint struct to checkpoints.rs
