+++ integration-tests/src/mpc_fixture/fixture_interface.rs
@@ -313,7 +313,7 @@ impl MpcFixtureNode {
     /// Returns the SyncUpdate response (IDs missing on this node).
     pub async fn sync(
         &self,
-        from: cait_sith::protocol::Participant,
+        from: Participant,
         triples: Vec<u64>,
         presignatures: Vec<u64>,
     ) -> SyncUpdate {
@@ -368,7 +368,7 @@ impl MpcFixtureNode {
     /// the peer from artifacts they don't have, pruning below threshold.
     pub async fn process_sync_response(
         &self,
-        peer: cait_sith::protocol::Participant,
+        peer: Participant,
         threshold: usize,
         response: &mpc_node::protocol::sync::SyncUpdate,
     ) {
