+++ integration-tests/tests/cases/mpc.rs
@@ -804,20 +804,16 @@ async fn test_sign_contention_5_nodes() {
 /// to keep per owner, then filters all nodes to that consistent set.
 /// Panics if any owner has fewer than `n` artifacts.
 fn truncate_per_owner<A: mpc_node::storage::protocol_storage::ProtocolArtifact>(
-    mut data: BTreeMap<
-        cait_sith::protocol::Participant,
-        BTreeMap<cait_sith::protocol::Participant, Vec<A>>,
-    >,
+    mut data: BTreeMap<Participant, BTreeMap<Participant, Vec<A>>>,
     n: usize,
-) -> BTreeMap<cait_sith::protocol::Participant, BTreeMap<cait_sith::protocol::Participant, Vec<A>>>
+) -> BTreeMap<Participant, BTreeMap<Participant, Vec<A>>>
 where
     A::Id: Ord,
 {
     use std::collections::BTreeSet;
 
     // Determine which IDs to keep per owner using the first node's ordering.
-    let mut keep_ids_per_owner: BTreeMap<cait_sith::protocol::Participant, BTreeSet<A::Id>> =
-        BTreeMap::new();
+    let mut keep_ids_per_owner: BTreeMap<Participant, BTreeSet<A::Id>> = BTreeMap::new();
     if let Some((first_node, first_owners)) = data.iter().next() {
         for (owner, artifacts) in first_owners {
             assert!(
@@ -848,11 +844,8 @@ where
 
 /// Filter artifact data to keep only artifacts that exist on ALL nodes.
 fn filter_artifacts_on_all_nodes<A: mpc_node::storage::protocol_storage::ProtocolArtifact>(
-    mut data: BTreeMap<
-        cait_sith::protocol::Participant,
-        BTreeMap<cait_sith::protocol::Participant, Vec<A>>,
-    >,
-) -> BTreeMap<cait_sith::protocol::Participant, BTreeMap<cait_sith::protocol::Participant, Vec<A>>>
+    mut data: BTreeMap<Participant, BTreeMap<Participant, Vec<A>>>,
+) -> BTreeMap<Participant, BTreeMap<Participant, Vec<A>>>
 where
     A::Id: Ord,
 {
